Add fallback apple semantic colors
