+++ packages/theming/apple-theme/src/appleTheme.ts
@@ -1,5 +1,207 @@
 import { Theme, Typography, Spacing, FontWeightValue, FontSize, FontSizes, Variants } from '@fluentui-react-native/theme-types';
-import { getStockApplePalette, getStockAppleDarkPalette } from './appleColors';
+import { ApplePalette, getStockApplePalette, getStockAppleDarkPalette } from './fluentAppleColors';
+
+// Instantiates the Palette with colors from FluentUI Apple and the platform semantic colors
+export function paletteFromAppleColors(p: ApplePalette): Palette {
+  return {
+    background: p.gray200,
+    bodyStandoutBackground: p.gray40,
+    bodyFrameBackground: p.blue10,
+    bodyFrameDivider: p.blue10,
+    bodyText: p.blue10,
+    bodyTextChecked: p.blue10,
+    subText: p.blue10,
+    bodyDivider: p.blue10,
+
+    disabledBackground: p.blue10,
+    disabledText: p.blue10,
+    disabledBodyText: p.blue10,
+    disabledSubtext: p.blue10,
+    disabledBodySubtext: p.blue10,
+
+    focusBorder: p.blue10,
+    variantBorder: p.blue10,
+    variantBorderHovered: p.blue10,
+    defaultStateBackground: p.blue10,
+
+    errorText: p.blue10,
+    warningText: p.blue10,
+    errorBackground: p.blue10,
+    blockingBackground: p.blue10,
+    warningBackground: p.blue10,
+    warningHighlight: p.blue10,
+    successBackground: p.blue10,
+
+    inputBorder: p.blue10,
+    inputBorderHovered: p.blue10,
+    inputBackground: p.blue10,
+    inputBackgroundChecked: p.blue10,
+    inputBackgroundCheckedHovered: p.blue10,
+    inputForegroundChecked: p.blue10,
+    inputFocusBorderAlt: p.blue10,
+    smallInputBorder: p.blue10,
+    inputText: p.blue10,
+    inputTextHovered: p.blue10,
+    inputPlaceholderText: p.blue10,
+
+    buttonBackgroundChecked: p.blue10,
+    buttonBackgroundHovered: p.blue10,
+    buttonBackgroundCheckedHovered: p.blue10,
+    buttonBackgroundPressed: p.blue10,
+    buttonBackgroundDisabled: p.blue10,
+    buttonText: p.blue10,
+    buttonTextHovered: p.blue10,
+    buttonTextChecked: p.blue10,
+    buttonTextCheckedHovered: p.blue10,
+    buttonTextPressed: p.blue10,
+    buttonTextDisabled: p.blue10,
+    buttonBorderDisabled: p.blue10,
+    buttonBorderFocused: p.blue10,
+
+    primaryButtonBackground: p.blue10,
+    primaryButtonBackgroundHovered: p.blue10,
+    primaryButtonBackgroundPressed: p.blue10,
+    primaryButtonBackgroundDisabled: p.blue10,
+    primaryButtonBorder: p.blue10,
+    primaryButtonBorderFocused: p.blue10,
+    primaryButtonText: p.blue10,
+    primaryButtonTextHovered: p.blue10,
+    primaryButtonTextPressed: p.blue10,
+    primaryButtonTextDisabled: p.blue10,
+
+    accentButtonBackground: p.blue10,
+    accentButtonText: p.blue10,
+
+    menuBackground: p.blue10,
+    menuDivider: p.blue10,
+    menuIcon: p.blue10,
+    menuHeader: p.blue10,
+    menuItemBackgroundHovered: p.blue10,
+    menuItemBackgroundPressed: p.blue10,
+    menuItemText: p.blue10,
+    menuItemTextHovered: p.blue10,
+
+    listBackground: p.blue10,
+    listText: p.blue10,
+    listItemBackgroundHovered: p.blue10,
+    listItemBackgroundChecked: p.blue10,
+    listItemBackgroundCheckedHovered: p.blue10,
+
+    listHeaderBackgroundHovered: p.blue10,
+    listHeaderBackgroundPressed: p.blue10,
+
+    actionLink: p.blue10,
+    actionLinkHovered: p.blue10,
+    link: p.blue10,
+    linkHovered: p.blue10,
+    linkPressed: p.blue10,
+
+    buttonBackground: p.blue10,
+    buttonContent: p.blue10,
+    buttonBorder: p.blue10,
+    buttonIcon: p.blue10,
+
+    buttonHoveredBackground: p.blue10,
+    buttonHoveredBorder: p.blue10,
+    buttonHoveredContent: p.blue10,
+    buttonHoveredIcon: p.blue10,
+
+    buttonFocusedBackground: p.blue10,
+    buttonFocusedBorder: p.blue10,
+    buttonFocusedContent: p.blue10,
+    buttonFocusedIcon: p.blue10,
+
+    buttonPressedBackground: p.blue10,
+    buttonPressedBorder: p.blue10,
+    buttonPressedContent: p.blue10,
+    buttonPressedIcon: p.blue10,
+
+    buttonDisabledBackground: p.blue10,
+    buttonDisabledBorder: p.blue10,
+    buttonDisabledContent: p.blue10,
+    buttonDisabledIcon: p.blue10,
+
+    ghostBackground: p.blue10,
+    ghostContent: p.blue10,
+    ghostBorder: p.blue10,
+    ghostIcon: p.blue10,
+
+    ghostHoveredBackground: p.blue10,
+    ghostHoveredBorder: p.blue10,
+    ghostHoveredContent: p.blue10,
+    ghostHoveredIcon: p.blue10,
+
+    ghostFocusedBackground: p.blue10,
+    ghostFocusedBorder: p.blue10,
+    ghostFocusedContent: p.blue10,
+    ghostFocusedIcon: p.blue10,
+
+    ghostPressedBackground: p.blue10,
+    ghostPressedBorder: p.blue10,
+    ghostPressedContent: p.blue10,
+    ghostPressedIcon: p.blue10,
+
+    ghostDisabledBackground: p.blue10,
+    ghostDisabledBorder: p.blue10,
+    ghostDisabledContent: p.blue10,
+    ghostDisabledIcon: p.blue10,
+
+    brandBackground: p.blue10,
+    brandContent: p.blue10,
+    brandBorder: p.blue10,
+    brandIcon: p.blue10,
+
+    brandHoveredBackground: p.blue10,
+    brandHoveredBorder: p.blue10,
+    brandHoveredContent: p.blue10,
+    brandHoveredIcon: p.blue10,
+
+    brandFocusedBackground: p.blue10,
+    brandFocusedBorder: p.blue10,
+    brandFocusedContent: p.blue10,
+    brandFocusedIcon: p.blue10,
+
+    brandPressedBackground: p.blue10,
+    brandPressedBorder: p.blue10,
+    brandPressedContent: p.blue10,
+    brandPressedIcon: p.blue10,
+
+    brandDisabledBackground: p.blue10,
+    brandDisabledBorder: p.blue10,
+    brandDisabledContent: p.blue10,
+    brandDisabledIcon: p.blue10,
+
+    buttonCheckedBackground: p.blue10,
+    buttonCheckedContent: p.blue10,
+    buttonCheckedHoveredBackground: p.blue10,
+    buttonCheckedHoveredContent: p.blue10,
+
+    brandCheckedBackground: p.blue10,
+    brandCheckedContent: p.blue10,
+    brandCheckedHoveredBackground: p.blue10,
+    brandCheckedHoveredContent: p.blue10,
+
+    ghostCheckedBackground: p.blue10,
+    ghostCheckedContent: p.blue10,
+    ghostCheckedHoveredBackground: p.blue10,
+    ghostCheckedHoveredContent: p.blue10,
+    ghostCheckedHoveredBorder: p.blue10,
+
+    ghostSecondaryContent: p.blue10,
+    ghostFocusedSecondaryContent: p.blue10,
+    ghostHoveredSecondaryContent: p.blue10,
+    ghostPressedSecondaryContent: p.blue10,
+
+    brandSecondaryContent: p.blue10,
+    brandFocusedSecondaryContent: p.blue10,
+    brandHoveredSecondaryContent: p.blue10,
+    brandPressedSecondaryContent: p.blue10,
+
+    buttonDisabledSecondaryContent: p.blue10,
+    buttonHoveredSecondaryContent: p.blue10,
+    buttonPressedSecondaryContent: p.blue10,
+  };
+}
 
 function _appleTypography(): Typography {
   const appleDict = {
